Removed core-data module
Removed the core-data module from the project and updated the app module to reflect this change.

Updated the theme to use dark
 color scheme and removed unused code.

Moved MainActivity and MainScreen to ui.main package.

Added edge-to-edge support for MainActivity.

Added README.md file with project description and module structure.

Moved AppNavHost to navigation package and added NavigationModule.